Add Model loading
Prior, only the first mesh in a file would be loaded.
This change adds Resource::LoadModel() which loads all of the
meshes in the ASSIMP scene, and sets up material properties
(for the default material) according to the aiMaterial per mesh.

This is just for convenience of defining and loading assets;
in engine, the Model is decomposed into its constituent
Mesh and MeshRenderer objects, with the MeshRenderer components
being added to the relevant Actor.

Also, various small bugfix and interface improvements.