Fix texture memory leak on scene switches
Calls glDeleteTextures() on textures in Resource::textureCache
on ClearCache().
Previously, these texture handles were simply lost with the data
still owned by OpenGL - including on the host and GPU.
This led to a memory leak.
This isn't necessary for the other caches; e.g. Mesh objects frees
the VAO etc in its destructor, and the destructor is called when
the containing GameInstance is destroyed.
    (should we store these Texture IDs in an object which frees
    the OpenGL memory when it is dropped from memory? Maybe).

This change also changes the GameInstance pointer from a shared_ptr
to a unique_ptr. This let me ensure the GameInstance is actually
destroyed when a new one replaces it, and that I didn't have any
other references to it that might have been causing the leak.
If we ever find some reason to run more than one GameInstance at
once, we can revert this.

Note this means that the Resource loading (and cacheing!) is
per scene, so loading times etc are longer between scenes.
We could certainly introduce some cache or resource pool across
scenes in the future if ever necessary.